+++ internal/server/arrow.go
@@ -247,7 +247,7 @@ func (h *hub) updateArrows(players map[int32]*tracked) {
 				h.splashPotion(players, a.dim, a.x, a.y, a.z, a.potion, a.lingering)
 			}
 			if a.explode > 0 { // ghast/wither fireball detonates on impact
-				h.explodeAt(players, a.x, a.y, a.z, a.explode+2, a.explode*4)
+				h.explodeIn(players, a.dim, a.x, a.y, a.z, a.explode+2, a.explode*4)
 			}
 			if a.loyalty > 0 { // a loyal trident returns after striking rather than vanishing
 				a.returning = true